change int to double

+++ CalcActionLog.cs
@@ -3,10 +3,11 @@ namespace AppSharp_6;
 internal class CalcActionLog
 {
     public CalcAction CalcAction {get; private set;}
-    public int CalcArgument {get; private set;}
-    public CalcActionLog(CalcAction calcAction, int calcArgument)
+    public double CalcArgument {get; private set;}
+    public CalcActionLog(CalcAction calcAction, double calcArgument)
     {
         CalcAction = calcAction;
         CalcArgument = calcArgument;
     }
+
 }
